+++ pkg/annotation/annotation.go
@@ -4,6 +4,7 @@ const (
 	Docs                    = "giantswarm.io/docs"
 	InstanceID              = "aws-operator.giantswarm.io/instance"
 	MachineDeploymentSubnet = "machine-deployment.giantswarm.io/subnet"
+	NodeTerminateUnhealthy  = "alpha.node.giantswarm.io/terminate-unhealthy"
 	AWSMetadata             = "alpha.giantswarm.io/aws-metadata-v2"
 	AWSSubnetSize           = "alpha.giantswarm.io/aws-subnet-size"
 )